refactor: move config into shared module
