+++ src/features/study/hooks/useStudyActions.spec.tsx
@@ -31,7 +31,7 @@ const mocks = vi.hoisted(() => {
   };
 });
 
-vi.mock("@/hooks/useConfig", () => ({
+vi.mock("@/shared/config/useConfig", () => ({
   useConfig: () => {
     if (mocks.state == null) throw new Error("Mock state is not initialized");
     return mocks.state.config;
