+++ src/pages/card-view/ui/CardViewPage.tsx
@@ -5,7 +5,7 @@ import * as C from "@/constant";
 import type { Card } from "@/entities/card";
 import type { Deck } from "@/entities/deck";
 import { useActions } from "@/hooks/useActions";
-import { useConfig } from "@/hooks/useConfig";
+import { useConfig } from "@/shared/config/useConfig";
 import { useRemoteCollections } from "@/hooks/useRemoteCollections";
 import { RemoteReadBoundary } from "@/shared/ui/remote-read-boundary";
 import { RouteFeedback } from "@/shared/ui/route-feedback";
